Message queue now work again. small improvements of DS set. (Some problems with Global Atom TRACE). Better screen borders handle

+++ user/queue.h
@@ -74,6 +74,6 @@ extern int QUEUE_FindMsg( MESSAGEQUEUE FAR * msgQueue, HWND hwnd,
                           int first, int last );
 extern void QUEUE_RemoveMsg( MESSAGEQUEUE FAR * msgQueue, int pos );
 extern void FAR hardware_event( WORD message, WORD wParam, LONG lParam,
-			    int xPos, int yPos, DWORD time, DWORD extraInfo );
+			    WORD xPos, WORD yPos, DWORD time, DWORD extraInfo );
 
 #endif  /* __WINE_QUEUE_H */
